+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Remove InputFile, it wasn't a real model
 
 ## [0.34.0]
 - Update Telegram API to 6.6: https://core.telegram.org/bots/api#march-9-2023